Modularised the Consumer and other Utils, status: Working
