+++ backend/.env.example
@@ -1,2 +1,2 @@
 IP_ADD=YOUR_IP:PORT
-GPU=NONE #NVIDIA OR NONE
+GPU=NONE #NVIDIA OR INTEGRATED
